some refactoring was made (https://github.com/OHDSI/WebAPI/issues/1220)
