+++ src/main/java/org/ohdsi/cohortcharacterization/design/BaseCriteriaFeatureImpl.java
@@ -2,11 +2,13 @@ package org.ohdsi.cohortcharacterization.design;
 
 import org.ohdsi.analysis.WithId;
 import org.ohdsi.analysis.cohortcharacterization.design.BaseCriteriaFeature;
+import org.ohdsi.analysis.cohortcharacterization.design.FeatureAnalysisAggregate;
 
 public abstract class BaseCriteriaFeatureImpl<T> implements BaseCriteriaFeature<T>, WithId<Long> {
 
 	private String name;
 	private Long id;
+    private FeatureAnalysisAggregateImpl aggregate;
 
 	@Override
 	public String getName() {
@@ -29,4 +31,13 @@ public abstract class BaseCriteriaFeatureImpl<T> implements BaseCriteriaFeature<
 
 		this.id = id;
 	}
+
+    @Override
+    public FeatureAnalysisAggregate getAggregate() {
+        return aggregate;
+    }
+
+    public void setAggregate(FeatureAnalysisAggregateImpl aggregate) {
+        this.aggregate = aggregate;
+    }
 }
