+++ src/main/java/org/ohdsi/cohortcharacterization/design/FeatureAnalysisWithCriteriaImpl.java
@@ -1,15 +1,14 @@
 package org.ohdsi.cohortcharacterization.design;
 
-import java.util.List;
-import org.ohdsi.analysis.cohortcharacterization.design.FeatureAnalysisAggregate;
 import org.ohdsi.analysis.cohortcharacterization.design.FeatureAnalysisWithCriteria;
 import org.ohdsi.circe.cohortdefinition.ConceptSet;
+
+import java.util.List;
 
 public abstract class FeatureAnalysisWithCriteriaImpl<T> extends FeatureAnalysisImpl<List<T>> implements FeatureAnalysisWithCriteria<T, Integer> {
 
 	private List<T> design;
 	private List<ConceptSet> conceptSets;
-	private FeatureAnalysisAggregateImpl aggregate;
 
 	@Override
 	public List<T> getDesign() {
@@ -31,16 +30,5 @@ public abstract class FeatureAnalysisWithCriteriaImpl<T> extends FeatureAnalysis
 	public void setConceptSets(List<ConceptSet> conceptSets) {
 
 		this.conceptSets = conceptSets;
-	}
-
-	@Override
-	public FeatureAnalysisAggregate getAggregate() {
-
-		return aggregate;
-	}
-
-	public void setAggregate(FeatureAnalysisAggregateImpl aggregate) {
-
-		this.aggregate = aggregate;
 	}
 }
